server/requirements.txt
flask>=2.0.1
onnxruntime>=1.14.0
numpy>=1.24.3
pillow>=9.5.0
flask-cors>=4.0.0
torch
torchvision
